user permissions added to delete item

+++ backend/src/resolvers/Mutation.js
@@ -50,8 +50,14 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     //  Find Item
-    const item = await ctx.db.query.item({ where }, `{id, title}`);
+    const item = await ctx.db.query.item({ where }, `{id, title, user {id}}`);
     // TODO: Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error('You do not have permission to do that!')
+    }
     // Delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
